+++ experimental/Pomdog.Experimental/Graphics/SpriteRenderer.cpp
@@ -262,14 +262,14 @@ void SpriteRenderer::Impl::DrawInstance(std::shared_ptr<Texture2D> const& textur
 	}
 
 	POMDOG_ASSERT(sprites.size() <= MaxBatchSize);
-	instanceVertices->SetData(sprites.data(), static_cast<std::uint32_t>(sprites.size()));
+	instanceVertices->SetData(sprites.data(), sprites.size());
 
 	graphicsContext->SetTexture(0, texture);
 	graphicsContext->SetVertexBuffers({planeVertices, instanceVertices});
 	graphicsContext->SetEffectPass(effectPass);
 	graphicsContext->SetConstantBuffers(constantBuffers);
 	graphicsContext->DrawIndexedInstanced(PrimitiveTopology::TriangleList,
-		planeIndices, planeIndices->IndexCount(), static_cast<std::uint32_t>(sprites.size()));
+		planeIndices, planeIndices->IndexCount(), sprites.size());
 }
 //-----------------------------------------------------------------------
 void SpriteRenderer::Impl::Draw(std::shared_ptr<Texture2D> const& texture, Matrix3x2 const& worldMatrix,
